[chore] : #6 SocialLoginTextDriver 컴포넌트로 변경
